% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netdemR-package.R
\docType{package}
\name{netdemR-package}
\alias{netdemR}
\alias{netdemR-package}
\title{Tools for analysis of Twitter data}
\description{
This package provides a series of functions that allow lab
members to to collect and analyze Twitter data, and to access the
datasets collected by the lab.
}
\author{
Pablo Barbera \email{pbarbera@usc.edu}
}

